Fix Empty response error.
So alot of users with home assistant have been reporting errors coming from WazeRouteCalculator.  These errors are essentially emtpy response or Internal Error which come directly from waze.  I've tracked the error down to one simple fact.  If you make a route request to the wrong server first, the subsequent requests get ignored or also fail.  This must be new to the API because it wasn't always like this.  So the fix is the use location based routing server requests.  Similar to how the COORD_SERVERS work.

I tested this in a build of home assistant with 11 separate routes.  3 in US using PRIVATE, TAXI, and MOTORCYCLE, 3 in EU using PRIVATE, TAXI, and MOTORCYCLE, 3 in IL using PRIVATE, TAXI, and MOTORCYCLE, and 2 others in IL and EU (random ones just for testing home assistant settings).  In this test calculate_all_routes getting performed twice a minute for a maximum of 1320 requests per hour. 

Prior to PR, roughly 776 errors per hour.

After PR, roughly 124 errors per hour.

The current 124 errors seem to plague Taxi routes in EU and Motorcycle routes in IL.  I'm not 100% sure what is causing that problem.  However, the private routes had about 4 errors for the hour among the 4 routes.